Move Level to a new file
